Translating headers to english
Renaming ComplementTemplates to ComplementaryTemplates

ComplementaryTemplates: Adding types to the parameters, and switching their order so that the optional ones come last.

ComplementaryTemplates: changing the docstring a little

Moving ComplementaryTemplates from the TemplateOperation subpackage into the TemplateGeneration subpackage.

Adding Zorandir to contributors.md

ExceptionTemplates: Changing the order of parameters so that the optional ones come last